requirements.txt
django
djangorestframework
django-heroku
waitress